runtime-rs: add generic support for running the VMM in non-root mode
This commit introduces generic support for running the VMM in rootless mode in runtime-rs:
1.Detect whether the VMM is running in rootless mode.
2.Before starting the VMM process, create a non-root user and launch the VMM with that user’s UID and GID; also add the KVM user's group ID to the VMM process's supplementary groups so the VMM process can access /dev/kvm.
3.Add the setup of the rootless directory located in the dir /run/user/<uid> directory, and modify some path variables to be functions that return the path with the rootless directory prefix when running in rootless mode.

Fixes: #11414

+++ src/runtime-rs/crates/resource/src/manager_inner.rs
@@ -505,10 +505,12 @@ impl ResourceManagerInner {
 
     pub async fn cleanup(&self) -> Result<()> {
         // clean up cgroup
-        self.cgroups_resource
-            .delete()
-            .await
-            .context("delete cgroup")?;
+        if !kata_types::rootless::is_rootless() {
+            self.cgroups_resource
+                .delete()
+                .await
+                .context("delete cgroup")?;
+        }
 
         // cleanup sandbox bind mounts: setup = false
         self.handle_sandbox_bindmounts(false)
